requirements.txt
python-telegram-bot[job-queue]==21.6
groq
aiomysql
sqlalchemy[asyncio]
apscheduler
python-dotenv
pytz
